// src/Pages/Modules/sale/delivery/Delivery.type.ts
interface DataType {
    key: string;
    deliveryNumber: string;
    deliveryDate: string;
    orderNumber: string;
    customer: string;
    deliveryType: string;
    deliveryCharge: string;
    deliveryChannel: string;
    channelCharge: string;
    consignmentId: string;
    deliveryStatus: string;
}

export type { DataType };